tests

+++ internal/models/accrual.go
@@ -1,7 +1,7 @@
 package models
 
 type Accrual struct {
-	Order   string  `json:"order"`
-	Status  string `json:"status"`
-	Accrual *int64 `json:"accrual,omitempty"`
+	Order   string   `json:"order"`
+	Status  string   `json:"status"`
+	Accrual *float64 `json:"accrual,omitempty"`
 }
