+++ internal/models/balance.go
@@ -2,6 +2,6 @@ package models
 
 type Balance struct {
 	//may be use float
-	Current   int64 `json:"current"`
-	Withdrawn int64 `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
