+++ internal/models/withdraw.go
@@ -1,6 +1,6 @@
 package models
 
 type Withdraw struct {
-	Order int64 `json:"order"`
-	Sum   int64 `json:"sum"`
+	Order int64   `json:"order"`
+	Sum   float64 `json:"sum"`
 }
