+++ internal/models/withdrawals.go
@@ -1,7 +1,7 @@
 package models
 
 type Withdrawals struct {
-	Order       int64  `json:"order"`
-	Sum         int64  `json:"sum"`
-	ProcessedAt string `json:"processed_at"`
+	Order       int64   `json:"order"`
+	Sum         float64 `json:"sum"`
+	ProcessedAt string  `json:"processed_at"`
 }
